Adding a new training area
Back at it. Going through yet another golang tutorial.

Every six months or so I pick up a golang tutorial and hack through to
somewhere around chapter three. Let's see if I go past that chapter with
this one, and keep at it for more than an hour or so :disappointed:.